# openapi.yaml
openapi: 3.0.3
info:
  title: Redstick Ventures Command Center API
  description: |
    Complete API for the Redstick Ventures Command Center platform.
    
    ## Authentication
    All API requests require a JWT token in the Authorization header:
    ```
    Authorization: Bearer <token>
    ```
    
    ## Rate Limiting
    - Authenticated: 1000 requests/minute
    - Unauthenticated: 20 requests/minute
    
    ## Base URLs
    - Production: https://api.redstick.vc
    - Staging: https://api-staging.redstick.vc
    - Development: http://localhost:3000/api
  version: 1.0.0
  contact:
    name: Redstick API Support
    email: api-support@redstick.vc
    url: https://docs.redstick.vc
  license:
    name: Proprietary
    url: https://redstick.vc/terms
servers:
  - url: https://api.redstick.vc/api/v1
    description: Production
  - url: https://api-staging.redstick.vc/api/v1
    description: Staging
  - url: http://localhost:3000/api/v1
    description: Development
security:
  - bearerAuth: []
paths:
...
  /deals:
    get:
      tags:
        - Deals
      summary: List all deals
      description: Get a paginated list of deals with filtering and sorting
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
        - name: stage
          in: query
          schema:
            type: string
            enum: [lead, qualified, pitch, duediligence, negotiation, closed, passed]
        - name: status
          in: query
          schema:
            type: string
            enum: [active, onhold, archived]
        - name: assignedTo
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: minAmount
          in: query
          schema:
            type: number
        - name: maxAmount
          in: query
          schema:
            type: number
        - name: fromDate
          in: query
          schema:
            type: string
            format: date-time
        - name: toDate
          in: query
          schema:
            type: string
            format: date-time
        - name: include
          in: query
          schema:
            type: string
            description: Comma-separated list of related data to include
      responses:
        '200':
          description: List of deals
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DealListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    post:
      tags:
        - Deals
      summary: Create a new deal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDealRequest'
      responses:
        '201':
          description: Deal created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DealResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/ConflictError'
  /deals/{id}:
    get:
      tags:
        - Deals
      summary: Get a specific deal
      parameters:
        - $ref: '#/components/parameters/DealIdParam'
        - name: include
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Deal details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DealDetailResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'

    patch:
      tags:
        - Deals
      summary: Update a deal
      parameters:
        - $ref: '#/components/parameters/DealIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDealRequest'
      responses:
        '200':
          description: Deal updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DealResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'

    delete:
      tags:
        - Deals
      summary: Delete/archive a deal
      parameters:
        - $ref: '#/components/parameters/DealIdParam'
        - name: permanent
          in: query
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Deal deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
  /users/{id}:
    get:
      tags:
        - Users
      summary: Get user by ID
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '403':
          $ref: '#/components/responses/ForbiddenError'

    patch:
      tags:
        - Users
      summary: Update user
      description: Admin only
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'

    delete:
      tags:
        - Users
      summary: Deactivate user
      description: Admin only
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      responses:
        '200':
          description: User deactivated

  # Dashboard
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
    SortByParam:
      name: sortBy
      in: query
      schema:
        type: string
        default: createdAt
    SortOrderParam:
      name: sortOrder
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: desc
    DealIdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Deal ID
...
    UserIdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: User ID
...
  schemas:
...
    SuccessResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
        data:
          type: object

    # Pagination
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
        hasNextPage:
          type: boolean
        hasPrevPage:
          type: boolean

    # User
    User:
      type: object
      required:
        - id
        - name
        - email
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
        roles:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, inactive, pending]
        avatar:
          type: string
          format: uri
        department:
          type: string
        title:
          type: string
        phone:
          type: string
        lastLoginAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    UserResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/User'
...
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        avatar:
          type: string
        department:
          type: string
        title:
          type: string
        phone:
          type: string

    # Authentication
...
    Deal:
      type: object
      required:
        - id
        - companyName
        - stage
      properties:
        id:
          type: string
        companyId:
          type: string
        companyName:
          type: string
        stage:
          type: string
          enum: [lead, qualified, pitch, duediligence, negotiation, closed, passed]
        status:
          type: string
          enum: [active, onhold, archived]
        amount:
          type: number
        description:
          type: string
        source:
          type: string
        sourceDetail:
          type: string
        probability:
          type: number
          minimum: 0
          maximum: 100
        expectedCloseDate:
          type: string
          format: date
        assignedTo:
          type: string
        assignee:
          $ref: '#/components/schemas/User'
        tags:
          type: array
          items:
            type: string
        customFields:
          type: object
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        createdBy:
          type: string
        lastActivityAt:
          type: string
          format: date-time
    DealResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Deal'
    DealListResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Deal'
            pagination:
              $ref: '#/components/schemas/Pagination'
    DealDetailResponse:
      allOf:
        - $ref: '#/components/schemas/DealResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                company:
                  $ref: '#/components/schemas/Company'
                activities:
                  type: array
                  items:
                    $ref: '#/components/schemas/Activity'
                documents:
                  type: array
                  items:
                    $ref: '#/components/schemas/File'
                notes:
                  type: array
                  items:
                    type: object
                tasks:
                  type: array
                  items:
                    type: object
    CreateDealRequest:
      type: object
      required:
        - companyName
        - stage
      properties:
        companyName:
          type: string
        companyId:
          type: string
        stage:
          type: string
          enum: [lead, qualified, pitch, duediligence, negotiation, closed, passed]
        status:
          type: string
          enum: [active, onhold, archived]
        amount:
          type: number
        description:
          type: string
        source:
          type: string
        sourceDetail:
          type: string
        probability:
          type: number
        expectedCloseDate:
          type: string
          format: date
        assignedTo:
          type: string
        tags:
          type: array
          items:
            type: string
        customFields:
          type: object
    UpdateDealRequest:
      allOf:
        - $ref: '#/components/schemas/CreateDealRequest'
      type: object

    # Company
    Company:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        website:
          type: string
          format: uri
        industry:
          type: string
        subIndustry:
          type: string
        stage:
          type: string
        foundedYear:
          type: integer
        employees:
          type: integer
        location:
          type: object
          properties:
            city:
              type: string
            state:
              type: string
            country:
              type: string
        socialLinks:
          type: object
          properties:
            linkedin:
              type: string
            twitter:
              type: string
        founders:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              title:
                type: string
              linkedin:
                type: string
        fundingHistory:
          type: array
          items:
            type: object
        metrics:
          type: object
          properties:
            revenue:
              type: number
            growthRate:
              type: number
            customers:
              type: integer
        dealCount:
          type: integer
        activeDealId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Activity:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [meeting, call, email, note, task, document, status_change]
        title:
          type: string
        description:
          type: string
        date:
          type: string
          format: date-time
        duration:
          type: integer
        participants:
          type: array
          items:
            type: string
        outcome:
          type: string
        nextSteps:
          type: string
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
...
    File:
      type: object
      properties:
        id:
          type: string
        filename:
          type: string
        originalName:
          type: string
        mimeType:
          type: string
        size:
          type: integer
        url:
          type: string
          format: uri
        thumbnailUrl:
          type: string
          format: uri
        uploadedBy:
          type: string
        createdAt:
          type: string
          format: date-time
...
    Error:
      type: object
      required:
        - success
        - error
        - code
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
        code:
          type: string
        details:
          type: object
        requestId:
          type: string
  responses:
    UnauthorizedError:
      description: Unauthorized - Invalid or missing token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: "Invalid or missing authentication token"
            code: "UNAUTHORIZED"
            requestId: "req_abc123"
    ForbiddenError:
      description: Forbidden - Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: "Insufficient permissions"
            code: "FORBIDDEN"
            requestId: "req_def456"
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: "Resource not found"
            code: "NOT_FOUND"
            requestId: "req_ghi789"
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: "Validation failed"
            code: "VALIDATION_ERROR"
            details:
              companyName: ["Company name is required"]
            requestId: "req_jkl012"
    ConflictError:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: "Deal already exists for this company"
            code: "DUPLICATE_RESOURCE"
            requestId: "req_mno345"
...